requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
python-dotenv==1.0.0
openai>=1.10.0,<2.0.0
requests>=2.25.0 